backend/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
PyJWT==2.8.0
google-cloud-vision==3.7.2
google-genai>=1.0.0
python-dotenv==1.0.1
Pillow==10.3.0
